fix: IDS, adding firewall rule to PSA and documentation (#650)
improving documentation and security controls
adding a firewall rule to ensure egress traffic towards private service access is allowed.
ADDITIONAL CONTEXT: when you deploy an IDS endpoint, a forwarding rule is created within the IP range assigned for private service access. To ensure IDS works, all GCE and GKE have to be able to send traffic to this ip. Traffic is replicated to this ip because of the packet mirroring policy that is attached to all resources specified by the filter.