fix: revert client-folder name to original value  (#712)
revert client-folder name to original value by removing the environment from the name.

Remember that in our scenario, the organization name identify the environment. If we want a client(app team) to see what environment he is working on (using the cloud console) we need to grant him organization viewer role. That role requires to be granted using the tier1 repo because that is where organization level roles are granted. We did not want to modify the tier1 repo every time a new client was onboarded so instead, we chose to embed the environment name in the client-folder name and have the folder viewer role granted in the tier2 repo.

The solution to this problem was to leverage nested groups.
1- organization viewer role is granted to a parent group which includes all of the client folder viewer group
2- folder viewer role is granted to the client folder viewer group which in turn is the parent group for all application team groups.

using this solution, remove the needs to have the environment within the client folder name as all application team will endup nested under the client folder viewer group (single nest) and the organization viewer group (double nest)